requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.1.1
django-date-extensions==3.0
django-filter==2.2.0
django-grappelli==2.13.3
django-split-settings==1.0.0
djangorestframework==3.10.3
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.5
packaging==19.2
pyparsing==2.4.5
pysolr==3.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
